test: split SDK protocol coverage
